Resolved four compiler warnings

+++ labwork3/src/codec3/dec_lossy_dct.cpp
@@ -113,8 +113,8 @@ int decode(args& cfg){
 		cv::Mat v(out.get_paduv_size(),CV_8U);
 
 		if(pred_mode==9){
-			for(uint by=0;by*predBlockSize<y.rows;by++){
-				for(uint bx=0;bx*predBlockSize<y.cols;bx++){
+			for(uint by=0;by*predBlockSize<(uint)y.rows;by++){
+				for(uint bx=0;bx*predBlockSize<(uint)y.cols;bx++){
 					uint bw_y=out.get_bsize_y().width;
 					uint bh_y=out.get_bsize_y().height;
 					uint bw_uv=out.get_bsize_uv().width;
